+++ R/spatvec.R
@@ -91,7 +91,7 @@ setMethod("dim", signature(x="SpatVector"),
 
 setMethod("as.data.frame", signature(x="SpatVector"), 
 	function(x, geom=NULL) {
-		d <- data.frame(x@ptr$getDF(), check.names=FALSE, fix.empty.names=FALSE, stringsAsFactors=FALSE)
+		d <- .getSpatDF(x@ptr$df)
 		# fix empty names 
 		colnames(d) <- x@ptr$names
 		if (!is.null(geom)) {